irl initial version
Former-commit-id: c3e641b7c5461d05cd74bca13288647ab76f6c95

+++ qamsi/cov_estimators/rl/env.py
@@ -33,7 +33,9 @@ class BanditEnvironment(gymnasium.Env):
 
         self.experiment_runner = experiment_runner
 
-        self.features = features
+        self.features = features.astype(
+            np.float32
+        )
 
         self.action_space = spaces.Box(low=0, high=1, dtype=np.float32)
 
@@ -141,7 +143,10 @@ class OptimalEnvironment(gymnasium.Env):
         super().__init__()
 
         self.experiment_runner = experiment_runner
-        self.features = features
+
+        self.features = features.astype(
+            np.float32
+        )
 
         self.optimal_rewards = vol_to_reward(optimal_vol).astype(
             np.float32
